+++ Core/Core.lua
@@ -126,6 +126,9 @@ function addon:OnPlayerLogin()
     
     -- Show welcome popup on first load
     self.WelcomePopup:Show()
+    
+    -- Show any pending migration notices
+    self.MigrationManager:Show()
 end
 
 function addon:OnPlayerLogout()